파일 계층 변경 및 db 환경설정
