+++ esg/.gitignore
@@ -1,5 +1,6 @@
 HELP.md
 .gradle
+
 build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
@@ -35,3 +36,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+## develop enviornment
+application.yml
