+++ esg/src/main/java/esgback/esg/Domain/Enum/State.java
@@ -1,4 +1,4 @@
-package Domain.Enum;
+package esgback.esg.Domain.Enum;
 
 public enum State {
     True, False
