+++ esg/src/main/java/esgback/esg/Domain/Notice/Notice.java
@@ -1,4 +1,4 @@
-package Domain.Notice;
+package esgback.esg.Domain.Notice;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
