+++ esg/src/main/java/esgback/esg/Domain/Purchase/PurchaseItem.java
@@ -1,27 +1,26 @@
-package Domain.ShoppingCart;
+package esgback.esg.Domain.Purchase;
 
-import Domain.Item.Item;
+import esgback.esg.Domain.Item.Item;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class ShoppingCartListedItem {
+public class PurchaseItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "shoppingCart_id")
-    private ShoppingCart shoppingCart;
+    @JoinColumn(name = "purchase_id")
+    private Purchase purchase;
 
     @OneToOne
     @JoinColumn(name = "item_id")
     private Item item;
 
-    private int shoppingCartListedItemQuantity;
-    private int totalPrice;
+    private int purchaseQuantity;
+    private int purchaseTotalPrice;
+    private int totalDiscountPrice;
 }
