+++ esg/src/main/java/esgback/esg/Domain/ShoppingCart/ShoppingCartListedItem.java
@@ -1,26 +1,27 @@
-package Domain.Purchase;
+package esgback.esg.Domain.ShoppingCart;
 
-import Domain.Item.Item;
+import esgback.esg.Domain.Item.Item;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PurchaseItem {
+@Getter
+public class ShoppingCartListedItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "purchase_id")
-    private Purchase purchase;
+    @JoinColumn(name = "shoppingCart_id")
+    private ShoppingCart shoppingCart;
 
     @OneToOne
     @JoinColumn(name = "item_id")
     private Item item;
 
-    private int purchaseQuantity;
-    private int purchaseTotalPrice;
-    private int totalDiscountPrice;
+    private int shoppingCartListedItemQuantity;
+    private int totalPrice;
 }
