+++ esg/src/main/java/esgback/esg/Domain/Wish/WishedItem.java
@@ -1,27 +1,22 @@
-package Domain.Reserve;
+package esgback.esg.Domain.Wish;
 
-import Domain.Item.Item;
+import esgback.esg.Domain.Item.Item;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.Date;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class ReservedItem {
+public class WishedItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "reserve_id")
-    private Reserve reserve;
 
     @OneToOne
     @JoinColumn(name = "item_id")
     private Item item;
 
-    private int quantity;
-    private int price;
+    private Date wishedItemDate;
 }
